Besic Level Serial Communication with UART
This project was created for easy communication between your FPGA and other devices includes UART interface like PC, MCU etc. It is very practical to import and use this in your project.

The project is consist of four verilog modules they are named as "UART.v",  "baudRateGenerator.v", "receiver.v", "transmitter.v".

The project has been tested both in simulation and on hardware.